// app/src/generated/models/Al_importbatchesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Al_importbatchesal_batchstatus = {
  120910730: 'Received',
  120910731: 'Validating',
  120910732: 'Completed'
} as const;
export type Al_importbatchesal_batchstatus = keyof typeof Al_importbatchesal_batchstatus;
export const Al_importbatchesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Al_importbatchesstatecode = keyof typeof Al_importbatchesstatecode;
export const Al_importbatchesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Al_importbatchesstatuscode = keyof typeof Al_importbatchesstatuscode;

export interface Al_importbatchesBase {
  al_batchstatus: Al_importbatchesal_batchstatus;
  al_exceptioncount?: number;
  al_importbatchcode: string;
  al_importbatchid: string;
  al_importedcount?: number;
  al_importedon?: string;
  al_name: string;
  al_source: string;
  al_totalrows?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Al_importbatchesstatecode;
  statuscode?: Al_importbatchesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Al_importbatches extends Al_importbatchesBase {
  al_batchstatusname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}